Modified the test to cover hold, abandon, and open_enrollment. These are tickets 21 and 94
